feat(asset): migrate asset-side money Float->Numeric(18,2)
Completes the money ledger unification started on the liability side
(#9 / b2c3d4e5f6a7 + c3d4e5f6a7b8). Migrates 9 Float money columns
across 3 tables to NUMERIC(18,2):

- assets.purchase_price / current_value / annual_maintenance_cost /
  target_daily_cost
- asset_valuations.value
- asset_lifecycle_events.sell_price / sell_fee

asset.interest_rate stays Float (percentage, mirrors liability.interest_rate).
Activity.amount stays Float (cross-entity display snapshot; see c3d4e5f6a7b8).

Money architecture: Decimal in compute / str on wire / float at aggregation
boundaries. AssetResponse/ValuationResponse/AssetSellResponse/
AssetLifecycleEventResponse money fields now serialize as 2-decimal str
via shared _coerce_money_str (reused from schemas/liability.py); Create/
Update/Sell/ValueUpdate inputs coerce to Decimal via _coerce_to_decimal.

Boundary fixes (Decimal/float mixed arithmetic raises TypeError):
- services/asset.py: compute_daily_cost / compute_return_rate / sell_asset
  arithmetic now float()-coerce at boundaries
- services/dashboard.py: ~10 ExchangeRateService.convert call sites +
  total_cost arithmetic float()-coerce (mirror liability-side fix)
- services/snapshot.py: convert(float(a.current_value or 0))
- services/projection.py: float(a.get("current_value"))
- services/whatif.py: float()-coerce asset dict money fields
- routers/assets.py: record_activity float()-coerce (Activity stays Float)
- routers/assets_analysis.py: total_held_cost + calculate_purchasing_power
  float()-coerce
- routers/export.py: str() asset money in CSV + JSON (json.dumps(Decimal)
  crashes; mirror review-fix 12f7b948)

dashboard schema (TopAssetItem etc.) stays float — Pydantic auto-coerces
Decimal->float for display aggregates. finance_coach_snapshot._money already
float(v)-coerces.

alembic d4e5f6a7b8c9 (down_revision c3d4e5f6a7b8) batch_alter 3 tables.

Frontend: Asset/AssetSellResponse/AssetValuation money fields number->string.
Dashboard-related interfaces (TopAssetItem/LowUsageItem/NewAssetItem) stay
number (backend dashboard schema is float).

Verification:
- backend 1232 passed / 1 skipped / 0 failed (full regression)
- domain 7 passed
- frontend 923/923 + typecheck clean
- ruff: no new errors (all pre-existing E711/E712)
- mypy: net -1 error on services/asset.py (compute_return_rate rewrite
  removed one; remaining line 88/262 pre-existing round() Any inference)

Tests: test_asset_response_returns_str_money (new str-serialization
regression) + updated test_assets/test_wishes assertions for str contract.

+++ server/apps/backend/app/models/asset_lifecycle_event.py
@@ -1,6 +1,7 @@
 from datetime import date, datetime
+from decimal import Decimal
 
-from sqlalchemy import BigInteger, Date, DateTime, Float, String, Text, func
+from sqlalchemy import BigInteger, Date, DateTime, Numeric, String, Text, func
 from sqlalchemy.orm import Mapped, mapped_column
 
 from apps.backend.app.database import Base
@@ -14,8 +15,8 @@ class AssetLifecycleEvent(Base):
     asset_id: Mapped[int] = mapped_column(BigInteger, nullable=False, index=True)
     event_type: Mapped[str] = mapped_column(String(20), nullable=False)  # 'sold' | 'retired'
     event_date: Mapped[date] = mapped_column(Date, nullable=False)
-    sell_price: Mapped[float | None] = mapped_column(Float, nullable=True)
-    sell_fee: Mapped[float | None] = mapped_column(Float, nullable=True)
+    sell_price: Mapped[Decimal | None] = mapped_column(Numeric(18, 2), nullable=True)
+    sell_fee: Mapped[Decimal | None] = mapped_column(Numeric(18, 2), nullable=True)
     sell_channel: Mapped[str | None] = mapped_column(String(100), nullable=True)
     notes: Mapped[str | None] = mapped_column(Text, nullable=True)
     created_at: Mapped[datetime] = mapped_column(DateTime, server_default=func.now())
